feat(permission): 实现批量权限管理和角色分配功能
- 修改 RoleChange 结构体的 targetEmail 字段为字符串数组类型
- 更新用户和角色仓库接口，支持批量用户ID和角色ID操作
- 添加 PermissionController 和 PermissionService 实现权限管理功能
- 创建权限管理相关的 DTO 数据传输对象
- 在服务器路由中注册权限和角色管理接口
- 实现批量角色分配和撤销的业务逻辑

+++ src/interfaces/server/dto/user.go
@@ -23,7 +23,7 @@ type BaseUserInfo struct {
 	AvatarUrl     string     `json:"avatar_url"`
 	QQ            string     `json:"qq"`
 	Rating        int        `json:"rating"`
-	Permission    int        `json:"permission"`
+	Permission    uint64     `json:"permission"`
 	RegisterTime  time.Time  `json:"register_time"`
 	LastLoginTime *time.Time `json:"last_login_time"`
 	LastLoginIp   *string    `json:"last_login_ip"`
@@ -42,7 +42,7 @@ func (b *BaseUserInfo) FromUserEntity(user *entity.User) *BaseUserInfo {
 	utils.ForEach(user.Roles, func(index int, role *entity.UserRole) {
 		perm.Merge(permission.Permission(role.Role.Permission))
 	})
-	b.Permission = int(perm)
+	b.Permission = uint64(perm)
 	b.RegisterTime = user.CreatedAt
 	if user.LastLoginTime.Valid {
 		b.LastLoginTime = &user.LastLoginTime.Time
